Move RegisterAvailable function from readers to writers
